+++ Proyecto_Programacion1/src/Arranque/Arranque.java
@@ -1,0 +1,11 @@
+
+package Arranque;
+
+import Vista.Inicio;
+
+
+public class Arranque {
+    public static void main(String[] args) {
+        Inicio ini = new Inicio();
+    }
+}
