+++ Proyecto_Programacion1/src/Clases/Countries.java
@@ -1,5 +1,5 @@
 
-package Vista;
+package Clases;
 
 import java.io.BufferedReader;
 import java.io.IOException;
