+++ Proyecto_Programacion1/src/Vista/Lista_paises.java
@@ -1,6 +1,7 @@
 
 package Vista;
 
+import Clases.Countries;
 import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
